finished first version

+++ jsecom_web/src/addToCart.js
@@ -1,29 +1,59 @@
 
 import { getCartProductFromLS } from "./getCartProducts.js";
-
+import { showToast } from "./showToast.js";
 import { updateCartValue } from "./updateCartValue.js";
 
 // -----------------------------------------------------
 // to get the cart data from localStorage
 // to update the cart value and also to get the data always ready from localStorage
 // --------------------------------------------------------
-// getCartProductFromLS();
+ getCartProductFromLS();
 
 // -----------------------------------------------------
 // to add the data into localStorage
 // --------------------------------------------------------
 export const addToCart = (event, id, stock) => {
-    console.log("kkk")
+    //console.log("kkk")
   let arrLocalStorageProduct = getCartProductFromLS();
 
   const currentProdElem = document.querySelector(`#card${id}`);
-  console.log(currentProdElem )
+  //console.log(currentProdElem )
   let quantity = currentProdElem.querySelector(".productQuantity").innerText;
   let price = currentProdElem.querySelector(".productPrice").innerText;
-    console.log(quantity, price);
+   // console.log(quantity, price);
+  
   price = price.replace("₹", "");
+   
+  let existingProd = arrLocalStorageProduct.find(
+    (curProd) => curProd.id === id
+  );
+
+  console.log(existingProd);
+    
+  if (existingProd && quantity > 1) {
+    quantity = Number(existingProd.quantity) + Number(quantity);
+    price = Number(price * quantity);
+    let updatedCart = { id, quantity, price };
+
+    updatedCart = arrLocalStorageProduct.map((curProd) => {
+      return curProd.id === id ? updatedCart : curProd;
+    });
+    console.log(updatedCart);
+
+    localStorage.setItem("cartProductLS", JSON.stringify(updatedCart));
+    //show toast when product added to the cart
+    showToast("add", id);
+  }
+
+  if (existingProd) {
+    // alert("bhai duplicate hai");
+    return false;
+  }
+
+  
+
   price = Number(price * quantity);
-   quantity = Number(quantity);
+  quantity = Number(quantity);
 
   arrLocalStorageProduct.push({ id, quantity, price });
   localStorage.setItem("cartProductLS", JSON.stringify(arrLocalStorageProduct));
@@ -31,8 +61,10 @@ export const addToCart = (event, id, stock) => {
   //update the cart button value
   updateCartValue(arrLocalStorageProduct);
 
-  
+  //show toast when product added to the cart
+  showToast("add", id);
 };
+
